# Speed up plot_optimizations.reorder() with Hilbert Curve