--- a/cli/requirements/requirements.txt
+++ b/cli/requirements/requirements.txt
@@ -9,4 +9,3 @@ plotink>=1.3.1
 pyserial>=3.5
 requests==2.25.1
 urllib3==1.26.5
-hilbertcurve==2.0.5