--- a/inkscape driver/plot_optimizations.py
+++ b/inkscape driver/plot_optimizations.py
@@ -50,7 +50,6 @@ These functions include:
 """
 
 import random
-import math
 import time
 import copy