Follow-up to https://github.com/evil-mad/axidraw/pull/112 speeding up `plot_optimizations.reorder()` for the same reasons.

This pull request introduces a spatial index to reduce needed comparisons with a parameter to trade off time and distance savings. I’ve included some initial results with time improvement over unoptimized and a comparison of pen-up distance reduction after reordering over unoptimized.

The pen-up distance improvements with optimization are generally comparable when `CURVE_DISTANCE = 100` but 80-90% faster to find. These time improvement are a bit less dramatic than the last PR but still meaningful.

The [Hilbert Curve library](https://pypi.org/project/hilbertcurve/) has NumPy as an expensive dependency, so it may be worth trying to port if this approach makes sense.

<table>
<tr>
<th>File</th><th>Paths</th><th><tt>CURVE_DISTANCE</tt></th><th>Time before #113</th><th>Distance savings before #113</th><th>Time after #113</th><th>Distance savings after #113</th><th>Time Improvement</th>
</tr>

<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-10-slateblue.svg">stroke-10-slateblue.svg</a></td><td>1</td><td>5</td><td>0.00 sec</td><td>0%</td><td>0.00 sec</td><td>—</td><td>—</td>
</tr>
<!--
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-11-navy.svg">stroke-11-navy.svg</a></td><td>3,424</td><td>5</td><td>0.00 sec</td><td>36%</td><td>0.00 sec</td><td>30%</td><td>—</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-11-navy.svg">stroke-11-navy.svg</a></td><td>3,424 (reverse)</td><td>5</td><td>0.00 sec</td><td>38%</td><td>0.00 sec</td><td>28%</td><td>—</td>
</tr>
-->
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-8-mediumblue.svg">stroke-8-mediumblue.svg</a></td><td>2,118</td><td>5</td><td>1.72 sec</td><td>88%</td><td>0.12 sec</td><td>64%</td><td>93%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-8-mediumblue.svg">stroke-8-mediumblue.svg</a></td><td>2,118 (reverse)</td><td>5</td><td>3.24 sec</td><td>95%</td><td>0.13 sec</td><td>63%</td><td>96%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-10-slateblue.svg">stroke-10-slateblue.svg</a></td><td>1</td><td>100</td><td>0.00 sec</td><td>0%</td><td>0.00 sec</td><td>—</td><td>—</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-11-navy.svg">stroke-11-navy.svg</a></td><td>3,424</td><td>100</td><td>4.63 sec</td><td>90%</td><td>0.52 sec</td><td>88%</td><td>89%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-11-navy.svg">stroke-11-navy.svg</a></td><td>3,424 (reverse)</td><td>100</td><td>9.07 sec</td><td>96%</td><td>0.68 sec</td><td>91%</td><td>93%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-8-mediumblue.svg">stroke-8-mediumblue.svg</a></td><td>2,118</td><td>100</td><td>1.72 sec</td><td>88%</td><td>0.32 sec</td><td>85%</td><td>81%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-8-mediumblue.svg">stroke-8-mediumblue.svg</a></td><td>2,118 (reverse)</td><td>100</td><td>3.24 sec</td><td>95%</td><td>0.39 sec</td><td>88%</td><td>88%</td>
</tr>
</table>