@oskay 2021-12-24
Neat! I was just starting to look at using the same R-tree from the previous PR as spatial index to improve the speed of a greedy search.

It's an interesting question as to what the right set of tradeoffs is on plot optimization. Ideally, we would pick a method that (typically) minimizes the total (plot optimization + plotting) time.  If optimizing is faster, but plotting takes longer, that won't necessarily be an improvement. 

As an aside on that topic, Paul Butler has a blog post that notes that TSP optimization past greedy/nearest-neighbor probably isn't a big win for plotting: https://nb.paulbutler.org/optimizing-plots-with-tsp-solver/