@migurski 2021-12-24
That’s a great post. Interesting that he proposes using an R-Tree. I was thinking about reusing the previous code, and decided not to because `reorder()` needs cover all points including very distant ones in contrast to `connect_nearby_ends()` which is just trying to find local hops. The Hilbert curve is not dissimilar to an R-Tree, but has the nice property of putting everything on a one-dimensional line. We could probably use `rtree.py` to generate a similar ordering of inputs.