@oskay 2021-12-24
I was looking at possibly trying a modified version of this approach: https://sigmodrecord.org/publications/sigmodRecord/9809/pdfs/290593.290596.pdf