@migurski 2021-12-25
I added total ordering to `rtree.py` and swapped out the Hilbert Curve in 63d0bd8. The results are probably just as good!

<table>
<tr>
<th>File</th><th>Paths</th><th><tt>CURVE_DISTANCE</tt></th><th>Time before optimizing</th><th>Distance savings before optimizing</th><th>Time after optimizing</th><th>Distance savings after optimizing</th><th>Time Improvement</th>
</tr>

<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-11-navy.svg">stroke-11-navy.svg</a></td><td>3,424</td><td>100</td><td>4.63 sec</td><td>90%</td><td>0.61 sec</td><td>88%</td><td>87%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-11-navy.svg">stroke-11-navy.svg</a></td><td>3,424 (reverse)</td><td>100</td><td>9.07 sec</td><td>96%</td><td>0.60 sec</td><td>92%</td><td>93%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-8-mediumblue.svg">stroke-8-mediumblue.svg</a></td><td>2,118</td><td>100</td><td>1.72 sec</td><td>88%</td><td>0.31 sec</td><td>87%</td><td>82%</td>
</tr>
<tr>
<td><a href="http://mike.teczno.com/img/axicli-examples-2021-12-19/stroke-8-mediumblue.svg">stroke-8-mediumblue.svg</a></td><td>2,118 (reverse)</td><td>100</td><td>3.24 sec</td><td>95%</td><td>0.44 sec</td><td>92%</td><td>86%</td>
</tr>
</table>