@migurski 2021-12-25
> Ideally, we would pick a method that (typically) minimizes the total (plot optimization + plotting) time. If optimizing is faster, but plotting takes longer, that won't necessarily be an improvement.

All the test times above are on my fast laptop. When I’m actually plotting I run `axicli` from a Raspberry Pi Zero 2 W where these big loops can take much longer. Would it be useful for me to benchmark there as well for this PR?