@oskay 2021-12-25
> Would it be useful for me to benchmark there as well for this PR?

I'm not sure that it's necessary; this is pretty close to being an "unqualified" improvement.