@migurski 2021-12-25
Some sample estimate results run on my laptop:

<table>
<tr>
<th>Estimate</th><th>Estimate time before optimizing</th><th>Print time before optimizing</th><th>Estimate time after optimizing</th><th>Print time after optimizing</th>
</tr>
<tr>
<td><tt>axicli --preview -T stroke-11-navy.svg</tt></td><td>1.77 sec</td><td>0:42:44</td><td>2.09 sec</td><td>0:42:44</td>
</tr>
<tr>
<td><tt>axicli --preview -G1 -T stroke-11-navy.svg</tt></td><td>2.36 sec</td><td>0:33:09</td><td>5.83 sec</td><td>0:32:45</td>
</tr>
<tr>
<td><tt>axicli --preview -G2 -T stroke-11-navy.svg</tt></td><td>1.72 sec</td><td>0:16:35</td><td>2.70 sec</td><td>0:16:17</td>
</tr>
<tr>
<td><tt>axicli --preview  -T stroke-8-mediumblue.svg</tt></td><td>13 sec</td><td>3:18:47</td><td>13 sec</td><td>3:18:47</td>
</tr>
<tr>
<td><tt>axicli --preview -G1 -T stroke-8-mediumblue.svg</tt></td><td>17 sec</td><td>3:07:05</td><td>226 sec</td><td>3:05:31</td>
</tr>
<tr>
<td><tt>axicli --preview -G2 -T stroke-8-mediumblue.svg</tt></td><td>17 sec</td><td>2:18:39</td><td>208 sec</td><td>2:13:42</td>
</tr>

</table>