@oskay 2021-12-25
We're using "optimizing" in a couple ways here. 😆 

If I'm reading this right, this is a slightly longer total (computation + print time) with this PR added -- on the last line, that's about five minutes longer plot time, but three minutes shorter computation time.