@migurski 2021-12-25
Good point, I’ve changed the headers to be less confusing. Your reading is right. I generally prefer to see the plot start more quickly and proceed more smoothly even if there’s a small time cost to prevent a dry pen tip. During the unpredictable pre-plotting time I scrape the pen tip with a post-it note to keep ink flowing, it’d be nice to not need to do that.