@oskay 2021-12-25
One of the things that I'm working on right now is a separate mode that outputs the optimized document, such that the optimization and plotting stages can be separated.