@migurski 2021-12-25
That would be amazing, I already do color separations before transferring files for plotting and it’d be nice to have them ready-to-spool directly to the AxiDraw.