@oskay 2021-12-28
It looks as though some intermediary performance can be achieved through a different choice of `CURVE_DISTANCE`.


Some sample estimate results run on my laptop,  with `CURVE_DISTANCE = 600`:
<table>
<tr>
<th>Estimate</th><th>Estimate time before #113</th><th>Print time before #113</th><th>Estimate time after #113</th><th>Print time after #113</th>
<tr>
<td><tt>axicli --preview  -T stroke-8-mediumblue.svg</tt></td><td>5.1 sec</td><td>3:18:47</td><td>5.1 sec</td><td>3:18:47</td>
</tr>
<tr>
<td><tt>axicli --preview -G1 -T stroke-8-mediumblue.svg</tt></td><td>1:11</td><td>3:05:31</td><td>11 sec</td><td>3:08:11</td>
</tr>
<tr>
<td><tt>axicli --preview -G2 -T stroke-8-mediumblue.svg</tt></td><td>1:12</td><td>2:13:42</td><td>11 sec</td><td>2:15:35</td>
</tr>

</table>

This saves about a minute in optimization, costs about two minutes in plot time. 

It may be better to consider a dynamic value of `CURVE_DISTANCE`, either based on the initial number of paths, or one that decreases with `len(spatial_index)`.  I've played with it a bit, and it looks like there are further tradeoffs available. However, in quick tests, I didn't see any see any choice that was unambiguously best