@j2kun 2021-12-29
Sorry for the random drive-by comment, but I'm a mathematician/programmer interested in cataloguing some use cases of Hilbert curves in practical settings. Can you confirm from the above discussion that the Hilbert curve ordering for your T-tree is _not_ going to be considered part of this PR? (In place of the z-ordering or "Morton" encoding, as it were?)