@oskay 2021-12-29
> Can you confirm from the above discussion that the Hilbert curve ordering for your T-tree is _not_ going to be considered part of this PR?

Not yet; sorry. I'm evaluating several approaches including some that do and do not include the Hilbert curve.