@oskay 2021-12-29
> It makes sense that a higher `CURVE_DISTANCE` would have this effect. Maybe something like `min(1000, len(spatial_index)/5)` to scale with the path count and still have an upper bound?

That's pretty close to what I was looking at.